(new-cli) added GitVersion.Core.Tester console
GitVersion.Core.Tester console will allow to host GitVersion.Core in a separate process.

For now the Calculation part will be exposed, the output, most probably will end up in the plugins, meaning it will not be possible to output from a separate host process. The consumer of the Core package will use the version variables as they consider fit.